Modifying auth model to split base URI
Generally I find it's best practice to be able to provide a separate base URI for an API. This allows, for example, a developer to switch between production and development applications. It also makes for less unwieldy API calls by letting you pass in just the partial path of the API call you're trying to make.

This also makes the Node implementation of EdgeGrid more in line with the Ruby implementation. I considered adding backwards compatibility (allowing a full URL to passed in via a `url` parameter on the request object) but I'd rather enforce this standard.