Adding a parser for .edgerc files
Most EdgeGrid authenticators parse `.edgerc` configuration files. This PR adds that functionality to this module.

Inside `api.js` I've added a `parseEdgerc()` function that handles the actual file parsing. I changed the API so that in addition to accepting four credentials as arguments, it also accepts an object with a `path` string pointing to an `.edgerc` file (or any file really, but .edgerc by convention), and a `group` that can be specified (since `.edgerc` files can contain multiple named groups of credentials).

I updated the Readme but didn't see an easy avenue for adding tests (since we're just testing the auth and not the API itself).